feat: homework-7 (50%)
