+++ index.html
@@ -57,6 +57,6 @@
 		<div class="wall-right"></div>
 	</div>
 
-	<script src="./script.js"></script>
+	<script src="./script.js" type="module"></script>
 </body>
 </html>
